+++ website/theme.config.tsx
@@ -73,7 +73,7 @@ export default defineConfig({
               children: 'Documentation',
             },
             {
-              href: 'https://the-guild.dev/blog',
+              href: 'https://the-guild.dev/graphql/hive/blog',
               icon: PencilIcon,
               children: 'Blog',
             },