+++ packages/plugins/typescript/typescript/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @graphql-codegen/typescript
 
+## 4.1.6
+
+### Patch Changes
+
+- Updated dependencies [[`f6909d1`](https://github.com/dotansimha/graphql-code-generator/commit/f6909d1797c15b79a0afb7ec089471763a485bfc)]:
+  - @graphql-codegen/visitor-plugin-common@5.8.0
+
 ## 4.1.5
 
 ### Patch Changes